Reorganized files; added package.json for SIOT JS lib
