+++ envsetup.sh
@@ -196,7 +196,7 @@ siot_setup_influx() {
 siot_protobuf() {
   echo "generating protobufs"
   protoc --proto_path=internal/pb internal/pb/*.proto --go_out=./ || return 1
-  protoc --proto_path=internal/pb internal/pb/*.proto --js_out=import_style=commonjs,binary:./internal/pb/ || return 1
+  protoc --proto_path=internal/pb internal/pb/*.proto --js_out=import_style=commonjs,binary:./frontend/lib/protobuf/ || return 1
 }
 
 siot_edge_run() {
