+++ frontend/lib/protobuf/point_pb.js
